pyproject.toml
[tool.poetry]
name = "fullstack-django-drf organic-farm-website"
version = "0.1.0"
description = ""
authors = ["Takib Yeasar <mdtakibyeasar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
django-tailwind = {extras = ["reload"], version = "^3.6.0"}
djangorestframework = "^3.14.0"
fontawesomefree = "^6.5.1"
unidecode = "^1.3.7"
setuptools = "^69.0.3"
django-countries = "^7.5.1"
stripe = "^7.11.0"
attrs = "^23.2.0"
python-dotenv = "^1.0.0"
django-browser-reload = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
